requirements.txt
# Algorithm Competition RAG System Dependencies
# Based on actual package versions from the virtual environment

# 核心AI依赖
zhipuai==2.1.5.20250801
zai-sdk==0.0.3.1

# 数据库依赖
neo4j==5.28.2

# 数值计算
numpy==2.2.6

# 系统监控
psutil==7.0.0

# Web框架（工程化版本使用）
Flask==3.1.2
Flask-SocketIO==5.5.1
eventlet==0.40.3

# 网络和HTTP
httpx==0.28.1
httpcore==1.0.9
certifi==2025.8.3
idna==3.10

# 异步支持
anyio==4.10.0
sniffio==1.3.1
exceptiongroup==1.3.0

# 类型注解
typing_extensions==4.14.1
typing-inspection==0.4.1
annotated-types==0.7.0

# 数据验证
pydantic==2.11.7
pydantic_core==2.33.2

# 缓存和工具
cachetools==6.1.0
bidict==0.23.1

# 模板引擎
Jinja2==3.1.6
MarkupSafe==3.0.2

# 安全
PyJWT==2.8.0
itsdangerous==2.2.0

# 时间处理
pytz==2025.2

# WebSocket支持
python-socketio==5.13.0
python-engineio==4.12.2
simple-websocket==1.1.0

# 网络协议
wsproto==1.2.0
h11==0.16.0
greenlet==3.2.4

# DNS解析
dnspython==2.8.0

# 信号处理
blinker==1.9.0

# 命令行工具
click==8.3.0

# 基础工具
setuptools==78.1.1
wheel==0.45.1